requirements.txt
torch>=1.8.0
numpy>=1.19.0
pyyaml>=5.4.0
tensorboard>=2.4.0
matplotlib>=3.3.0
tqdm>=4.50.0
scipy>=1.6.0
pytest>=6.2.0